% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_private_methods.R
\name{load_config}
\alias{load_config}
\title{Load Default Configuration}
\usage{
load_config()
}
\value{
List of default parameters
}
\description{
Initialize the default analysis parameters for the RNA-seq pipeline.
}
\keyword{internal}
