% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BulkRNASeqAnalysis_class.R,
%   R/03_task_functions.R
\name{perform_MAplot_task}
\alias{perform_MAplot_task}
\title{生成MA图 - 任务函数}
\usage{
perform_MAplot_task(self, ...)

perform_MAplot_task(self, ...)
}
\description{
生成MA图 - 任务函数

生成MA图 - 任务函数
}
